Also use list of installed RPMs for generating hash
As I noticed the other day Fedora 37's container image actually updated -
which means we're likely now carrying packages in caches that don't need to
be there anymore. The fastest way of detecting this is to simply include
the list of RPMs in our hash, since there doesn't seem to be any way to get
the image ID of the container.

Call rpm directly as well, since this is much faster then dnf.